+++ compiler/frontend/ast_exp_extension.ml
@@ -60,7 +60,7 @@ let handle_extension e (self : Bs_ast_mapper.mapper)
   | "re" ->
     Exp.constraint_ ~loc
       (Ast_exp_handle_external.handle_raw ~kind:Raw_re loc payload)
-      (Ast_comb.to_js_re_type loc)
+      (Ast_comb.to_regexp_type loc)
   | "external" -> (
     Location.deprecated loc
       "%external is deprecated, use %raw or regular FFI syntax instead.";