+++ compiler/frontend/ast_literal.ml
@@ -70,8 +70,7 @@ module Lid = struct
   (* FIXME: Use primitive module *)
   let js_null_undefined : t = Ldot (Lident "Js", "null_undefined")
 
-  (* FIXME: Use primitive module *)
-  let js_re_id : t = Ldot (Ldot (Lident "Js", "Re"), "t")
+  let regexp_id : t = Ldot (Lident "Stdlib_RegExp", "t")
 end
 
 module No_loc = struct